Renamed the pipeline file
